feat: clean seo
